Eliminate browser DOM module
